Split SSE Actix adapter tests
Move the crate-root Actix adapter wire-contract checks out of the BDD
scenario file so `tests/sse_wire_contract_bdd.rs` stays below the 400-line
limit while preserving the same integration coverage.
